move conditions
