+++ src/packages/core/section/conditions/section-user-permission.condition.ts
@@ -1,10 +1,7 @@
+import type { UmbSectionUserPermissionConditionConfig } from './types.js';
 import { UMB_CURRENT_USER_CONTEXT } from '@umbraco-cms/backoffice/current-user';
 import { UmbControllerBase } from '@umbraco-cms/backoffice/class-api';
-import type {
-	UmbConditionConfigBase,
-	UmbConditionControllerArguments,
-	UmbExtensionCondition,
-} from '@umbraco-cms/backoffice/extension-api';
+import type { UmbConditionControllerArguments, UmbExtensionCondition } from '@umbraco-cms/backoffice/extension-api';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 
 export class UmbSectionUserPermissionCondition extends UmbControllerBase implements UmbExtensionCondition {
@@ -31,12 +28,9 @@ export class UmbSectionUserPermissionCondition extends UmbControllerBase impleme
 	}
 }
 
-export type UmbSectionUserPermissionConditionConfig = UmbConditionConfigBase<'Umb.Condition.SectionUserPermission'> & {
-	/**
-	 *
-	 *
-	 * @example
-	 * "Umb.Section.Content"
-	 */
-	match: string;
+export const manifest = {
+	type: 'condition',
+	name: 'Section User Permission Condition',
+	alias: 'Umb.Condition.SectionUserPermission',
+	api: UmbSectionUserPermissionCondition,
 };
