+++ src/packages/documents/documents/user-permissions/conditions/manifests.ts
